refactor: Code cleaning (visual studio).

+++ src/PixelDust.Game/Worlding/World/PWorld.Elements.cs
@@ -11,7 +11,7 @@ namespace PixelDust.Core.Worlding
     {
         public void InstantiateElement<T>(Vector2Int pos) where T : PElement
         {
-            InstantiateElement(pos, (uint)elementDatabase.GetIdOfElementType<T>());
+            InstantiateElement(pos, elementDatabase.GetIdOfElementType<T>());
         }
         public void InstantiateElement(Vector2Int pos, uint id)
         {
@@ -23,7 +23,7 @@ namespace PixelDust.Core.Worlding
         }
         public bool TryInstantiateElement<T>(Vector2Int pos) where T : PElement
         {
-            return TryInstantiateElement(pos, (uint)elementDatabase.GetIdOfElementType<T>());
+            return TryInstantiateElement(pos, elementDatabase.GetIdOfElementType<T>());
         }
         public bool TryInstantiateElement(Vector2Int pos, uint id)
         {
